+++ js/form/validation-form.js
@@ -1,10 +1,10 @@
-const uploadFormElement = document.querySelector('.img-upload__form');
-const hashtagElement = uploadFormElement.querySelector('.text__hashtags');
-const textDescriptionElement = uploadFormElement.querySelector('.text__description');
-
 const LIMIT_HASHTAG = 5;
 const patternHashtag = /^#[a-zа-яё0-9]{1,19}$/i;
 const LIMIT_DESCRIPTION_LENGTH = 140;
+
+const uploadFormElement = document.querySelector('.img-upload__form');
+const hashtagElement = uploadFormElement.querySelector('.text__hashtags');
+const textDescriptionElement = uploadFormElement.querySelector('.text__description');
 
 const pristine = new Pristine(uploadFormElement, {
   classTo: 'img-upload__field-wrapper',
