+++ js/gallery/gallery.js
@@ -1,29 +1,31 @@
-import { getRenderingThumbnail } from './rendering-thumbnail.js';
-import { openBigPicture } from './rendering-big-picture.js';
-import { getData } from '../api.js';
+// import { getRenderingThumbnail } from './rendering-thumbnail.js';
+// import { openBigPicture } from './rendering-big-picture.js';
+// import { getData } from '../api.js';
 
-getData()
-  .then((data) => {
-    getRenderingThumbnail(data);
+// const picteresElement = document.querySelector('.pictures');
 
-    document.querySelector('.pictures').addEventListener('click', (evt) => {
-      const parent = evt.target.parentNode;
+// const TIME_CLEAR = 5000;
 
-      if (parent?.dataset?.id) {
-        evt.preventDefault();
-        const profile = data.find((item) => item.id.toString() === parent.dataset.id);
+// getData()
+//   .then((data) => {
+//     getRenderingThumbnail(data);
+//     picteresElement.addEventListener('click', (evt) => {
+//       const targetElement = evt.target.closest('.picture');
 
-        if (profile) {
-          openBigPicture(profile);
-        }
-      }
-    });
-  })
-  .catch(() => {
-    const errorTemplate = document.querySelector('#data-error').content.querySelector('.data-error');
-    document.body.append(errorTemplate);
+//       if (targetElement?.dataset?.id) {
+//         evt.preventDefault();
+//         const photo = data.find((item) => item.id.toString() === targetElement.dataset.id);
 
-    setTimeout(() => {
-      document.body.removeChild(document.querySelector('.data-error'));
-    }, 5000);
-  });
+//         if (photo) {
+//           openBigPicture(photo);
+//         }
+//       }
+//     });
+//   })
+//   .catch(() => {
+//     const errorTemplate = document.querySelector('#data-error').content.querySelector('.data-error');
+//     document.body.append(errorTemplate);
+//     setTimeout(() => {
+//       document.body.removeChild(document.querySelector('.data-error'));
+//     }, TIME_CLEAR);
+//   });
